de-10-ai-data-middleware/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
openai==1.51.0
aiofiles==24.1.0
pymysql==1.1.1
